fix: 워크스페이스별 트랜잭션 분리
